batch_bo/pyproject.toml
[project]
name = "batch-bo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "botorch>=0.12.0",
    "gpjax>=0.9.3",
    "matplotlib>=3.9.2",
    "poli-core>=1.1.0",
]
